Replace mixpanel with analytics-client
Also makes this module browser-only.
NodeJS support is being moved to analytics-client module.

Change-type: major
