+++ package.json
@@ -4,9 +4,9 @@
   "description": "Balena events logging util",
   "main": "build/balena-event-log.js",
   "dependencies": {
+    "analytics-client": "^0.6.2",
     "bluebird": "^3.4.7",
     "lodash": "^4.0.0",
-    "resin-mixpanel-client": "^2.7.0",
     "resin-universal-ga": "^1.2.2",
     "resin-universal-gosquared": "^0.2.0"
   },
@@ -27,10 +27,8 @@
   "scripts": {
     "bundle": "browserify src/balena-event-log.js -o build/bundle.js",
     "lint": "balena-lint --typescript --fix -e js src",
-    "test-node": "mocha \"./tests/**/*.js\"",
-    "test-browser": "karma start",
     "pretest": "npm run lint && npm run build",
-    "test": "npm run test-node && npm run test-browser",
+    "test": "karma start",
     "build": "npx tsc",
     "prepack": "npm run build"
   },
