--- src/adaptors/mixpanel.js
@@ -1,87 +1,0 @@
-var pick = require('lodash/pick');
-
-var MixpanelClient = require('resin-mixpanel-client');
-
-var ONE_TIME_USER_FIELDS = ['$created'];
-
-var UPDATE_USER_FIELDS = [
-	'$email',
-	'$name',
-	'hasPasswordSet',
-	'iat',
-	'id',
-	'permissions',
-	'public_key',
-	'username',
-];
-
-var getMixpanelUser = function (userData) {
-	var mixpanelUser = Object.assign(
-		{
-			$email: userData.email,
-			$name: userData.username,
-		},
-		userData,
-	);
-	return {
-		oneTime: pick(mixpanelUser, ONE_TIME_USER_FIELDS),
-		update: pick(mixpanelUser, UPDATE_USER_FIELDS),
-	};
-};
-
-module.exports = function (options) {
-	var debug = options.debug;
-	var token = options.mixpanelToken;
-	var mixpanelOptions = options.mixpanelHost
-		? {
-				api_host: options.mixpanelHost,
-				decide_host: options.mixpanelHost,
-		  }
-		: {};
-
-	if (!token) {
-		if (debug) {
-			console.warn('`mixpanelToken` is not set, Mixpanel tracking is disabled');
-		}
-		return null;
-	}
-
-	var mixpanel = MixpanelClient(token, mixpanelOptions);
-
-	return {
-		login: function (user, deviceIds) {
-			if (!user) return;
-			var methodName = user.$created ? 'signup' : 'login';
-			var mixpanelUser = getMixpanelUser(user);
-
-			return mixpanel[methodName](user.username, deviceIds)
-				.then(function () {
-					// Calling this also ensures that the auto-tracked properties
-					// ($os, $browser, $browser_version, $initial_referrer, $initial_referring_domain)
-					// are collected and sent
-					return mixpanel.setUser(mixpanelUser.update);
-				})
-				.then(function () {
-					return mixpanel.setUserOnce(mixpanelUser.oneTime);
-				});
-		},
-		logout: function () {
-			return mixpanel.logout();
-		},
-		track: function (prefix, type, data) {
-			return mixpanel.track('[' + prefix + '] ' + type, data);
-		},
-		getDistinctId: function () {
-			return { mixpanel: mixpanel.getDistinctId() };
-		},
-		identify: function (ids) {
-			var mixpanelId = ids['mixpanel'];
-
-			if (!mixpanelId) {
-				return null;
-			}
-
-			return mixpanel.identify(mixpanelId);
-		},
-	};
-};
